drop vague information about AOL
this pr drops the vague information about AOL completely.
it seems as if IMAP/SMPT works in general without OAuth2,
https://help.aol.com/articles/how-do-i-use-other-email-applications-to-send-and-receive-my-aol-mail
and also on https://github.com/deltachat/provider-db/issues/91 ,
the tests with thunderbird seems to work.

i did not test AOL at all, however, saying, it is BROKEN
is probably wrong.

so, before saying something that relies on very vague information,
i think it is better to drop the record completely.

i think this was also the gist in saying "drop" in
https://github.com/deltachat/provider-db/issues/91