clarify what happens when no servers are defined
it was a bit unclear to me (or i forgot it),
what happens when no servers are defined.

therefore, i added two explaining lines to the README.md

i suspected, that in this case, autoconfig/autodiscover are skipped -
which may have led to a worsening when for a provider a provider-entry is defined.

however, in fact, this is not the case.

the rust-core does not differ between missing provider-entries
or provider-entries without servers -
get_offline_autoconfig() just returns None in both cases.

for accessing additional information from a provider-entry,
there is an additional function, get_provider_info()