riseup.net: include domain into the username
I have checked by configuring testing account
manually with username being a full email
instead of just the local part.

This is how it is specified in
<https://autoconfig.riseup.net/mail/config-v1.1.xml>

I also tested manually with openssl-s_client
that IMAP server at mail.riseup.net:993 accepts username
both with and without the domain name.
Domain name is however not ignored,
it is impossible to login as "user@example.org"
if your email address is "user@riseup.net",
but using just "user" is fine.